feat(OPS-11): Fixes in order placement functionality.

+++ src/main/java/org/viators/orderprocessingsystem/product/ProductT.java
@@ -30,5 +30,5 @@ public class ProductT extends BaseEntity {
     private CategoryEnum category;
 
     @Column(name = "stock_quantity")
-    private BigDecimal stockQuantity;
+    private Long stockQuantity;
 }
